Chore/#181 separate history state change enums and coordinate extensions
