+++ talklat/talklat/Sources/Views/HistoryInfoItemView.swift
@@ -283,20 +283,6 @@ struct HistoryInfoItemView: View {
     }
 }
 
-
-extension MKCoordinateRegion: Equatable {
-    public static func == (lhs: MKCoordinateRegion, rhs: MKCoordinateRegion) -> Bool {
-        if lhs.center.latitude == rhs.center.latitude,
-           lhs.center.longitude ==  rhs.center.longitude,
-           lhs.span.latitudeDelta == rhs.span.latitudeDelta,
-           lhs.span.longitudeDelta == rhs.span.longitudeDelta {
-            return true
-        } else {
-            return false
-        }
-    }
-}
-
 #Preview {
     HistoryInfoItemView(
         historyInfoStore: TKHistoryInfoStore(),
